// __mock__/mongoose.model.d.ts
export declare const mockModel: {
    find: jest.Mock<any, any, any>;
    skip: jest.Mock<any, any, any>;
    limit: jest.Mock<any, any, any>;
    sort: jest.Mock<any, any, any>;
    populate: jest.Mock<any, any, any>;
    lean: jest.Mock<any, any, any>;
    exec: jest.Mock<any, any, any>;
    countDocuments: jest.Mock<any, any, any>;
    aggregate: jest.Mock<any, any, any>;
};